Adds UE4 Runtime References fTo Generated Modules Projects

+++ Managed/PluginInstaller/PluginInstaller/Program.cs
@@ -227,7 +227,14 @@ namespace PluginInstaller
         {
             if ((overwrite || !File.Exists(destFileName)) && File.Exists(sourceFileName))
             {
-                File.Copy(sourceFileName, destFileName, overwrite);
+                try
+                {
+                    File.Copy(sourceFileName, destFileName, overwrite);
+                }
+                catch
+                {
+                    Console.WriteLine("Failed to copy to '{0}'", destFileName);
+                }
             }
         }
 
@@ -438,26 +445,6 @@ namespace PluginInstaller
         {
             try
             {
-                string baseMicrosoftKeyPath = @"SOFTWARE\WOW6432Node\Microsoft";
-                string visualStudioRegistryKeyPath = baseMicrosoftKeyPath + @"\VisualStudio\SxS\VS7";
-
-                //Try Obtaining the VS version of MSBuild First
-                using (RegistryKey key = Registry.LocalMachine.OpenSubKey(visualStudioRegistryKeyPath))
-                {
-                    if (key != null)
-                    {
-                        string path = key.GetValue("15.0") as string;
-                        if (!string.IsNullOrEmpty(path))
-                        {
-                            path = Path.Combine(path, "MSBuild", "15.0", "Bin", "msbuild.exe");
-                            if (File.Exists(path))
-                            {
-                                return path;
-                            }
-                        }
-                    }
-                }
-
                 using (RegistryKey key = Registry.LocalMachine.OpenSubKey(@"SOFTWARE\Microsoft\MSBUILD\ToolsVersions\4.0"))
                 {
                     string path = key.GetValue("MSBuildToolsPath") as string;
@@ -577,6 +564,7 @@ namespace PluginInstaller
             bool shippingBuild = keyValues.ContainsKey("shipping");
             bool x86Build = keyValues.ContainsKey("x86");
             bool skipCopy = keyValues.ContainsKey("nocopy");
+            bool skipCleanup = keyValues.ContainsKey("noclean");
 
             string pluginName = "USharp";
             //string targetPrefix = "UE4Editor";
@@ -667,10 +655,17 @@ namespace PluginInstaller
                     string[] copyDirs = { "Binaries", "Intermediate" };
                     foreach (string dir in copyDirs)
                     {
-                        if(!Directory.Exists(Path.Combine(outputDir, dir)))
+                        if (Directory.Exists(Path.Combine(outputDir, dir)))
                         {
                             CopyFilesRecursive(new DirectoryInfo(Path.Combine(outputDir, dir)),
-                            new DirectoryInfo(Path.Combine(enginePluginDir, dir)), true);
+                                new DirectoryInfo(Path.Combine(enginePluginDir, dir)), true);
+
+                            // Also copy to the local path if outside of the engine?
+                            //if (!isInsideEngineFolder)
+                            //{
+                            //    CopyFilesRecursive(new DirectoryInfo(Path.Combine(outputDir, dir)),
+                            //        new DirectoryInfo(Path.Combine(localPluginDir, dir)), true);
+                            //}
                         }
                     }
 
@@ -699,15 +694,18 @@ namespace PluginInstaller
                 {
                 }
 
-                try
+                if (!skipCleanup)
                 {
-                    if (!Directory.Exists(outputDir))
+                    try
                     {
-                        Directory.Delete(outputDir);
+                        if (!Directory.Exists(outputDir))
+                        {
+                            Directory.Delete(outputDir);
+                        }
                     }
-                }
-                catch
-                {
+                    catch
+                    {
+                    }
                 }
             }
         }
